+++ src/modules/NativeCASNativeAdViewComponent.ts
@@ -15,13 +15,13 @@
  */
 
 import type { ColorValue, HostComponent, ViewProps } from 'react-native';
-import type { Int32 } from 'react-native/Libraries/Types/CodegenTypes';
+import type { Float, Int32 } from 'react-native/Libraries/Types/CodegenTypes';
 import codegenNativeComponent from 'react-native/Libraries/Utilities/codegenNativeComponent';
 
 export interface NativeProps extends ViewProps {
   instanceId: Int32;
-  width: Int32;
-  height: Int32;
+  width: Float;
+  height: Float;
   backgroundColor?: ColorValue;
   primaryColor?: ColorValue;
   primaryTextColor?: ColorValue;
